Split guest-bio parser coverage
Move the guest-biography BDD parser scenario into its own integration
test file so the general `parse_xml` behaviour module stays below the
repository file-size limit.
